parsefloat

+++ main.js
@@ -18,8 +18,8 @@ if (calcChoice == 1) {
 
   var operation = prompt("+", "-", "/", "*");
 
-  num1 = parseInt (num1);
-  num2 = parseInt (num2);
+  num1 = parseFloat (num1);
+  num2 = parseFloat (num2);
 
   // return reseult to user
     if (operation == "+"){
